implemented feedback: moved generateCode function to services
